requirements.txt
PySide6>=6.0.0
Pillow>=9.0.0
imagehash>=4.3.0

